Adds an import function via API endpoint
Moves the import script to a server route for manual triggering.

This change replaces the previous file watcher implementation with
an API endpoint that can be called to initiate the import process.
It enhances control over the import process and allows for integration
with a user interface.  The importer page now contains a button to trigger this import.