Tracks export status and prevents re-exporting
Adds a `last_exported_at` field to the image table to track when an image was last exported.

Updates the exporter to only export images that have been updated since their last export or have never been exported.

Adds `needsExport` flag to exporter API to indicate if an image needs to be re-exported.

Updates the triage page to provide visual feedback on restoring images.